allocateNetwork: fix network sandbox not cleaned up on failure
The defer function was checking for the local `err` variable, not
on the error that was returned by the function. As a result, the
sandbox would never be cleaned up for containers that used "none"
networking, and a failiure occured during setup.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 1a5611a32cfdace3eaafc7ec2b9f2c60b2458028)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>